% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compPI.R
\name{compPI}
\alias{compPI}
\title{Computation of prediction intervals}
\usage{
compPI(xNew, x0, L, L_max = 35L, theta, lambda, gradient, type, alpha, I,
  withSolve = TRUE)
}
\arguments{
\item{xNew}{values of influential variable}

\item{x0}{observed values of influential variable}

\item{L}{integer value indicating the number of events to be predicted (used when xNew not given)}

\item{L_max}{integer value indicating the maximum number of events in Poisson process when xNew is not given}

\item{theta}{numeric vector of length four with link function's parameters}

\item{gradient}{gradient of link function w.r.t. parameters}

\item{type}{if link or gradient are not given, there is collection of implented link functions and gradients, see \code{\link{linkfun}} and \code{\link{gradLambda}}}

\item{alpha}{confidence niveau in (0,1)}

\item{I}{information matrix}

\item{withSolve}{logical value indicating if inverting of information matrix with solve() or ginv() from MASS}

\item{link}{link function for the negative (!) logarithmized mean of the exponential distribution}
}
\value{
list with prediction interval bounds and corresponding quantiles of hypoexponential
}
\description{
Computation of prediction intervals
}

