% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeResiduals.R
\name{computeResiduals}
\alias{computeResiduals}
\title{Residuals of fitted median of exponential distribution used for data depth}
\usage{
computeResiduals(theta, x, y, lambda, type)
}
\arguments{
\item{theta}{numeric vector of length two or four with link function's parameters}

\item{x}{values of influental variable for the link function}

\item{y}{values of dependent variable}

\item{lambda}{[\code{function(theta, x)}]\cr
link function for exponential distribution}

\item{type}{[\code{integer}]\cr
if link function is not given a collection of given link function is available, see \code{\link{linkfun}}}

\item{gradient}{[\code{function(x, theta, ...)}]\cr
gradient of link function}
}
\value{
vector of residuals
}
\description{
Residuals of fitted median of exponential distribution used for data depth
}

