% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeStress.R
\name{computeStress}
\alias{computeStress}
\title{Computation of load sharing rule}
\usage{
computeStress(initial, nJumps, L_max = 35)
}
\arguments{
\item{initial}{Initial stress value when no components are damaged}

\item{nJumps}{Number of failures for which the load sharing should be computed}

\item{L_max}{Number of components in the system}
}
\value{
Vector of initial stress and the stress values with respect to the load sharing rule
}
\description{
Computation of load sharing rule
}

