% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confsetToInterval.R
\name{confsetToInterval}
\alias{confsetToInterval}
\title{Computes the quantiles of the hypoexponential distribution for all parameter values in a confidence set}
\usage{
confsetToInterval(cs, lambda, test.s, alpha)
}
\arguments{
\item{cs}{Matrix of parameter values for the link function of the hypoexponential distribution.
Each row of the matrix is one set of parameters}

\item{lambda}{function(x, theta) link function to link the value of x to the parameter of the hypoexponential distribution}

\item{test.s}{Vector of x values for the link function}

\item{alpha}{The alpha/2 and the 1 - alpha/2 quantiles of the hypoexponential distribution are computed}
}
\value{
A list of all intervals for the parameters in the confidence set
}
\description{
Computes the quantiles of the hypoexponential distribution for all parameter values in a confidence set
}

