% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{Computation of prediction intervals}
\usage{
delta(stresses, deltat, truss, start, toPred, link, gradient, type,
  plot = FALSE, xlim, alpha = 0.05, addTrue = TRUE, withSolve = TRUE,
  ...)
}
\description{
Computation of prediction intervals
}

