% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estI.R
\name{estI}
\alias{estI}
\title{Estimation of information matrix}
\usage{
estI(x, theta)
}
\arguments{
\item{x}{values of influental variable for the link function}

\item{theta}{numeric vector of length four with link function's parameters}
}
\value{
estimated information matrix
}
\description{
Estimation of information matrix
}

