% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estML.R
\name{estML}
\alias{estML}
\title{Maximum likelihood estimation of parameters in link function for the exponential distribution}
\usage{
estML(x, t, start, ...)
}
\arguments{
\item{x}{values of influental variable for the link function}

\item{t}{values of dependent variable}

\item{start}{starting solution for optim()}

\item{...}{further paraments passed to optim()}
}
\value{
optimal solution for the parameter in the link function in optimum and estimated
        information matrix in I
}
\description{
Maximum likelihood estimation of parameters in link function for the exponential distribution
}

