% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateCandidates.R
\name{generateCandidates}
\alias{generateCandidates}
\title{Grid generation around a given two-dimensional point needed for confidence sets}
\usage{
generateCandidates(center, range = c(0.5, 1.5), steps = center * 0.05)
}
\arguments{
\item{center}{center of the grid, a two-dimensional point}

\item{range}{two-dimensional vector specifiying the width of the region around the center}

\item{steps}{numeric value indicating the step width of the grid}
}
\value{
grid around \code{center} using \code{expand.grid}
}
\description{
Candidate generation for the confidence sets around a given a center with specified range and step width
}

