% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFit.R
\name{getFit}
\alias{getFit}
\title{Computation of fitted values for new data}
\usage{
getFit(newdata, theta, plot = FALSE, median = FALSE)
}
\arguments{
\item{newdata}{values of influental variable}

\item{theta}{parameter vector for link function}

\item{plot}{logical value indicating whether the fitted values should be added to an existing plot}

\item{median}{logical value indicating whether the median (TRUE) oder the expectation should be fitted (FALSE, default)}
}
\value{
vector of fitted values
}
\description{
Computation of fitted values for new data
}

