% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getQuantiles.R
\name{getQuantiles}
\alias{getQuantiles}
\title{Computation of the alpha/2 and the 1 - alpha/2 quantiles of the hypoexponential distribution for given rates of exponential distributions}
\usage{
getQuantiles(rates, alpha)
}
\arguments{
\item{rates}{Vector of rates of exponential distributions}

\item{alpha}{The alpha/2 and the 1 - alpha/2 quantiles of the hypoexponential distribution are computed}
}
\value{
alpha/2 and the 1 - alpha/2 quantiles of the hypoexponential distribution
}
\description{
Computation of the alpha/2 and the 1 - alpha/2 quantiles of the hypoexponential distribution for given rates of exponential distributions
}

