% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradBterm.R
\name{gradBterm}
\alias{gradBterm}
\title{Collection of gradient functions of H function for delta-method}
\usage{
gradBterm(type)
}
\arguments{
\item{type}{integer (1-6) indicating the desired link function, see \code{\link{linkfun}}}
}
\value{
gradient of H function at b with respect to theta
}
\description{
Collection of gradient functions of H function for delta-method
}

