% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradLambda.R
\name{gradLambda}
\alias{gradLambda}
\title{Gradient of link function with respect to parameter theta}
\usage{
gradLambda(x, theta)
}
\arguments{
\item{x}{value of influental variable for the link function}

\item{theta}{numeric vector of length four with link function's parameters}
}
\value{
gradient of link function at x with respect to theta
}
\description{
Gradient of link function with respect to parameter theta
}

