% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inInterval.R
\name{inInterval}
\alias{inInterval}
\title{Check if entries of a vector are contained in an interval}
\usage{
inInterval(x, int)
}
\arguments{
\item{x}{Vector with entries which should be checked}

\item{int}{Interval given by a vector of two length 2, the first entry is the lower bound, the second entry is the upper bound}
}
\value{
Boolean vector indicating if the elements of x are contained in the interval
}
\description{
Check if entries of a vector are contained in an interval
}

