% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{Link function for continous four parameter model}
\usage{
lambda(x, theta)
}
\arguments{
\item{x}{value of influental variable}

\item{theta}{vector of length four, parameter of link function}
}
\value{
value of the link function
}
\description{
Link function for continous four parameter model
}

