% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lastColumn.R
\name{lastColumn}
\alias{lastColumn}
\title{Returns last column of a matrix}
\usage{
lastColumn(m)
}
\arguments{
\item{m}{matrix}
}
\value{
last column of m
}
\description{
Returns last column of a matrix
}

