% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lastList.R
\name{lastList}
\alias{lastList}
\title{Returns last element of a list}
\usage{
lastList(l)
}
\arguments{
\item{l}{List}
}
\value{
last element of list l
}
\description{
Returns last element of a list
}

