% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkfun.R
\name{linkfun}
\alias{linkfun}
\title{A collection of link functions for the estimation of S-N-Curves.}
\usage{
linkfun(type)
}
\arguments{
\item{type}{integer (1-6) indicating the desired link function}
}
\value{
function(theta, x)
}
\description{
A collection of link functions for the estimation of S-N-Curves.
}

