% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plots data for a S-N-curve}
\usage{
plotData(x, t, ...)
}
\arguments{
\item{x}{stress values for the x-axis}

\item{t}{values for y-axis}

\item{...}{further arguments passed to plot()}
}
\value{
plot of x and log10(t)
}
\description{
Plots data for a S-N-curve
}

