% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotJumps.R
\name{plotJumps}
\alias{plotJumps}
\title{Plot event times}
\usage{
plotJumps(jumptimes, log = FALSE, ...)
}
\arguments{
\item{jumptimes}{Event times to be plotted}

\item{log}{logical value indicating whether to plot log10(jumptimes) or not, default FALSE}

\item{...}{further arguments passed to plot()}
}
\value{
Plot of event times
}
\description{
Plot event times
}

