% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predBeam.R
\name{predBeam}
\alias{predBeam}
\title{Compute prediction intervals for state dependent Poisson process for beam experiments}
\usage{
predBeam(stresses, deltat, truss, start, toPred, plot = FALSE,
  plotControl = list(stairs = TRUE, xlim = c(0, 0), legend = TRUE),
  withSolve = TRUE)
}
\arguments{
\item{stresses}{list of values of the influental variable for indepedent fatigue experiments.}

\item{deltat}{list of waiting times for indepedent fatigue experiments}

\item{truss}{index of list entry in stresses and deltat for which the prediction should be made}

\item{start}{starting value for the optimization, length four}

\item{toPred}{integer value indicating the number of events to be predicted}

\item{plot}{logical value indicating whether the prediction intervals should be plotted or not}

\item{plotControl}{list with arguments for plot: stairs (logical, plot as step function), xlim, legend (logical, generate legend)}

\item{withSolve}{logical value indicating whether the information matrix should be inverted with solve() (TRUE) or ginv() from MASS package (FALSE)}
}
\description{
Compute prediction intervals for state dependent Poisson process for beam experiments
}

