\name{predfat-package}
\alias{predfat-package}
\alias{predfat}
\docType{package}
\title{
Prediction methods for fatigue experiments
}
\description{
Prediction methods for fatigue experiments
}
\details{
\tabular{ll}{
Package: \tab predfat\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-17\cr
License: \tab What license is it under?\cr
}

}
\author{
Sebastian Szugat

Maintainer: <szugat@statistik.tu-dortmund.de>
}
\references{

\keyword{ package }
\seealso{


}
\examples{

}
