% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predInt.R
\name{predInt}
\alias{predInt}
\title{Prediction interval for the hypoexponential distribution based on delta method}
\usage{
predInt(rates, newdata, alpha, I, link, gradient, theta)
}
\arguments{
\item{rates}{Rates of the exponential distributions whose sum are hypoexponential distributed}

\item{newdata}{Values for the link function which belong to the vector rates}

\item{alpha}{The alpha/2 and 1 - alpha/2 quantiles of the hypoexponential distribution are computed}

\item{I}{information matrix}

\item{gradient}{function(x,y,theta,lambda,gradient) gradient of the link function}

\item{theta}{Vector with parameters of the link function}
}
\value{
Prediction interval for jumps at newdata using the delta method and the hypoexponential distribution
}
\description{
Prediction interval for the hypoexponential distribution based on delta method
}

