% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qhypoexp.R
\name{qhypoexp}
\alias{qhypoexp}
\title{p-quantile from the hypo-exponential distribution}
\usage{
qhypoexp(p, rate, interval = c(0, 10^10))
}
\arguments{
\item{p}{vector of probabilities}

\item{rate}{vector of (unique) rates}

\item{interval}{search interval in which \code{uniroot} searches for the quantile}
}
\value{
p-quantile from the hypo-exponential distribution
}
\description{
The p-quantile from the hypo-exponential distribution is computed implicitly by computing the root of H(x) = F(x) - p,
where F is the distribution function of the hypo-exponential distribution
}

