% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qs.R
\name{qs}
\alias{qs}
\title{Solver for quadratic equations a*x^2 + b*x + c = 0}
\usage{
qs(a, b, c)
}
\arguments{
\item{a}{real number}

\item{b}{real number}

\item{c}{real number}
}
\value{
solution of the quadratic equation a*x^2 + b*x + c = 0
}
\description{
Solver for quadratic equations a*x^2 + b*x + c = 0
}
\author{
Ted Harding, \email{Ted.Harding@nessie.mcc.ac.uk}
}

