% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/range2.R
\name{range2}
\alias{range2}
\title{Computes the range of a vector}
\usage{
range2(x, ...)
}
\arguments{
\item{x}{vector for which the range should be computed}

\item{...}{parameters passed to base function \code{range}}
}
\value{
returns the range of a vector, but in contrast to base R it returns the difference of \code{range}
}
\description{
Computes the range of a vector
}

