% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeResiduals.R
\name{computeResiduals}
\alias{computeResiduals}
\title{Residuals of fitted median of exponential distribution used for data depth}
\usage{
computeResiduals(theta, x, y)
}
\arguments{
\item{theta}{numeric vector of length two or four with link function's parameters}

\item{x}{values of influental variable for the link function}

\item{y}{values of dependent variable}
}
\value{
vector of residuals
}
\description{
Residuals of fitted median of exponential distribution used for data depth
}

