% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estML.R
\name{estML}
\alias{estML}
\title{Maximum likelihood estimation of parameters in link function for the exponential distribution}
\usage{
estML(x, t, delta, link, type = 1, start, ...)
}
\arguments{
\item{x}{values of influental variable for the link function}

\item{t}{values of dependent variable}

\item{delta}{vector indicating of length length(t) indicating if the respective observation in t is completely observed (1) or censored (0)}

\item{link}{a function(theta, log) used in the likelihood which is \eqn{L(\theta) = \sum t * \lambda(x) - log(\lambda(x))}}

\item{type}{[\code{integer}]\cr
if link function is not given a collection of given link function is available, see \code{\link{linkfun}}}

\item{start}{starting solution for optim()}

\item{...}{further paraments passed to optim()}
}
\value{
optimal solution for the parameter in the link function in optimum and estimated
        information matrix in I
}
\description{
Maximum likelihood estimation of parameters in link function for the exponential distribution
}

