% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFit.R
\name{getFit}
\alias{getFit}
\title{Computation of fitted values for new data}
\usage{
getFit(newdata, theta, lambda, type, plot = FALSE, median = FALSE)
}
\arguments{
\item{newdata}{values of influental variable}

\item{theta}{parameter vector for link function}

\item{lambda}{[\code{function(theta, x)}]\cr
link function for exponential distribution}

\item{type}{[\code{integer}]\cr
if link function is not given a collection of given link function is available, see \code{\link{linkfun}}}

\item{plot}{logical value indicating whether the fitted values should be added to an existing plot}

\item{median}{logical value indicating whether the median (TRUE) oder the expectation should be fitted (FALSE, default)}
}
\value{
vector of fitted values
}
\description{
Computation of fitted values for new data
}

