% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradLambda.R
\name{gradLambda}
\alias{gradLambda}
\title{Collection of gradient functions of link function for exponential distribution}
\usage{
gradLambda(type = 4)
}
\arguments{
\item{type}{integer (1-6) indicating the desired link function, see \code{\link{linkfun}}}
}
\value{
gradient of link function at x with respect to theta
}
\description{
Collection of gradient functions of link function for exponential distribution
}

