% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qhypoexp.R
\name{qhypoexp}
\alias{qhypoexp}
\title{Implicit computation of the p-quantile from the hypoexponential distribution}
\usage{
qhypoexp(p, rates, interval)
}
\arguments{
\item{p}{probability for the quantile in [0,1]}

\item{rates}{vector of rates of hypoexponential distribution}

\item{interval}{search interval in which uniroot() searches for the quantile}
}
\value{
p-quantile from the hypoexponential distribution
}
\description{
Implicit computation of the p-quantile from the hypoexponential distribution
}

